Bar Sela

+++ Engine/src/engine/engine.java
@@ -14,7 +14,7 @@ public interface engine{
     Information targetsInformation();
     Information specificTargetInformation(String name) throws Exception;
     Information findAPathBetweenTwoTargets(String t1,String t2 , Dependence d) throws Exception;
-    List<Information> runTask(float time,boolean random,float success,float warning,boolean keepLastRun) throws Exception;
+    List<Information> runTask(int time, boolean random, float success, float warning,boolean keepLastRun) throws Exception;
     Information circuitDetection(String name)throws Exception;
 
     void exit();
